requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.4
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-import-export==4.3.9
django-querycount==0.8.3
djangorestframework==3.16.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
